+++ source/common/http/async_client_impl.h
@@ -100,6 +100,7 @@ private:
     const std::string& maxAge() const override { return EMPTY_STRING; };
     const absl::optional<bool>& allowCredentials() const override { return allow_credentials_; };
     bool enabled() const override { return false; };
+    const std::string& runtimeKey() const override { return EMPTY_STRING; };
 
     static const std::list<std::string> allow_origin_;
     static const std::list<std::regex> allow_origin_regex_;