+++ source/common/router/config_impl.h
@@ -105,6 +105,7 @@ public:
   const std::string& maxAge() const override { return max_age_; };
   const absl::optional<bool>& allowCredentials() const override { return allow_credentials_; };
   bool enabled() const override { return enabled_; };
+  const std::string& runtimeKey() const override { return runtime_key_; };
 
 private:
   std::list<std::string> allow_origin_;
@@ -115,6 +116,7 @@ private:
   std::string max_age_{};
   absl::optional<bool> allow_credentials_{};
   bool enabled_;
+  std::string runtime_key_;
 };
 
 class ConfigImpl;