+++ source/extensions/filters/http/cors/config.cc
@@ -13,7 +13,7 @@ Http::FilterFactoryCb
 CorsFilterFactory::createFilter(const std::string& stats_prefix,
                                 Server::Configuration::FactoryContext& context) {
   CorsFilterConfigSharedPtr config =
-      std::make_shared<CorsFilterConfig>(stats_prefix, context.scope());
+      std::make_shared<CorsFilterConfig>(stats_prefix, context.scope(), context.runtime());
   return [config](Http::FilterChainFactoryCallbacks& callbacks) -> void {
     callbacks.addStreamFilter(std::make_shared<CorsFilter>(config));
   };