+++ test/mocks/router/mocks.h
@@ -53,6 +53,7 @@ public:
   const std::string& maxAge() const override { return max_age_; };
   const absl::optional<bool>& allowCredentials() const override { return allow_credentials_; };
   bool enabled() const override { return enabled_; };
+  const std::string& runtimeKey() const override { return runtime_key_; };
 
   std::list<std::string> allow_origin_{};
   std::list<std::regex> allow_origin_regex_{};
@@ -62,6 +63,7 @@ public:
   std::string max_age_{};
   absl::optional<bool> allow_credentials_{};
   bool enabled_{false};
+  std::string runtime_key_{};
 };
 
 class TestRetryPolicy : public RetryPolicy {